fix map offer
