+++ src/store/action.ts
@@ -1,3 +1,4 @@
 import { createAction } from '@reduxjs/toolkit';
 
 export const changeCity = createAction<{city: string}>('changeCity');
+
